+++ packages/orq-rc/RELEASES.md
@@ -6779,3 +6779,13 @@ Based on:
 - [typescript v4.3.0-rc.64] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.64 - packages/orq-rc
+
+## 2026-02-13 13:51:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.715.0 (2.822.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.65] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.65] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.65 - packages/orq-rc